requirements.txt
Flask==2.3.3
python-telegram-bot==20.4
python-dotenv==1.0.0
gunicorn==21.2.0